pyproject.toml
[project]
name = "mysystemdown"
version = "0.1.0"
description = "MySystemDown is a health check system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.1",
    "django-tenants>=3.10.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "psycopg2>=2.9.11",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "redis[hiredis]>=7.1.1",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=40.4.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
pythonpath = ["src"]